Align them so they are very similar
